Fix Caculator LLM

+++ src/main/java/com/sep/vox/application/port/input/command/ConsumeQuotaCommand.java
@@ -1,5 +1,6 @@
 package com.sep.vox.application.port.input.command;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 import com.sep.vox.domain.model.subscription.QuotaType;
@@ -8,7 +9,7 @@ public record ConsumeQuotaCommand(
     UUID subscriptionId,
     UUID examSessionId,
     QuotaType quotaType,
-    Integer amount,
+    BigDecimal amount,
     UUID userId
 ) {
 }
