+++ src/main/java/com/sep/vox/interfaces/rest/dto/request/PlanQuotaItemRequest.java
@@ -8,7 +8,7 @@ import jakarta.validation.constraints.NotNull;
 
 public record PlanQuotaItemRequest(
     @NotNull QuotaType quotaType,
-    @NotNull Integer includedQuantity,
+    @NotNull BigDecimal includedQuantity,
     @NotNull BigDecimal tokenUnitPrice
 ) {
 }
